chore: Organise repo (#237)
chore: Update pubspec to point to icons correctly
