+++ example/pubspec.yaml
@@ -11,22 +11,11 @@ dependencies:
   cached_network_image: ^3.3.1
   flutter:
     sdk: flutter
-  flutter_svg: ^2.0.10+1
   go_router: ^14.1.2
-  google_fonts: ^6.2.1
-  markdown_widget: ^2.3.2+6
   path_drawing: ^1.0.1
-  shared_preferences: ^2.2.2
-  url_launcher: ^6.2.6
   zeta_flutter:
     path: ../
 
-dev_dependencies:
-  flutter_test:
-    sdk: flutter
-  widgetbook: ^3.10.0
-
-  http: any
 flutter:
   uses-material-design: true
   assets:
