+++ lib/src/components/avatars/avatar.dart
@@ -66,6 +66,7 @@ enum ZetaAvatarSize {
 /// An avatar is a visual representation of a user or entity.
 ///
 /// It is recommended to use [ZetaAvatar] with [ZetaAvatarBadge] for status and notification badges, but any widget can be used.
+///
 /// {@category Components}
 ///
 /// Figma: https://www.figma.com/file/JesXQFLaPJLc1BdBM4sisI/%F0%9F%A6%93-ZDS---Components?type=design&node-id=20816-388
@@ -435,6 +436,8 @@ enum ZetaAvatarBadgeType {
 // TODO(UX-1138): Can this be refactored to use ZetaIndicator?
 
 /// Badge component used with [ZetaAvatar] as either [ZetaAvatar.upperBadge] or [ZetaAvatar.lowerBadge].
+///
+/// {@category Components}
 ///
 /// Sizes and styles are managed by the parent [ZetaAvatar].
 class ZetaAvatarBadge extends StatelessWidget {
