+++ lib/src/components/bottom sheets/bottom_sheet.dart
@@ -108,6 +108,8 @@ class ZetaBottomSheet extends ZetaStatelessWidget {
 /// Function to show [ZetaBottomSheet].
 ///
 /// Uses [showModalBottomSheet] for functionality, but with Zeta styling and simplified functionality.
+///
+/// {@category Components}
 Future<T?> showZetaBottomSheet<T>({
   required BuildContext context,
   String? title,
