+++ lib/src/components/buttons/button.dart
@@ -2,6 +2,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 import '../../../zeta_flutter.dart';
+import 'button_style.dart';
 
 /// Zeta Button
 /// {@category Components}
