+++ lib/src/components/chat_item/chat_item.dart
@@ -383,6 +383,8 @@ enum _ZetaSlidableActionType {
 }
 
 /// Slidable action widget for [ZetaChatItem].
+///
+/// {@category Components}
 class ZetaSlidableAction extends StatelessWidget {
   /// Constructs a [ZetaSlidableAction].
   const ZetaSlidableAction({
