+++ lib/src/components/checkbox/checkbox.dart
@@ -9,6 +9,7 @@ import '../../../zeta_flutter.dart';
 /// Widgets that use a checkbox should listen for the [onChanged] callback and
 /// rebuild the checkbox with a new [value] to update the visual appearance of
 /// the checkbox.
+///
 /// {@category Components}
 ///
 /// Figma: https://www.figma.com/file/JesXQFLaPJLc1BdBM4sisI/%F0%9F%A6%93-ZDS---Components?node-id=21510-54003
@@ -95,6 +96,8 @@ class ZetaCheckbox extends FormField<bool> {
 }
 
 /// [FormFieldState] for [ZetaCheckbox].
+///
+/// {@category Components}
 class ZetaCheckboxFormFieldState extends FormFieldState<bool> {
   @override
   ZetaCheckbox get widget => super.widget as ZetaCheckbox;
