+++ lib/src/components/chips/chip.dart
@@ -3,14 +3,16 @@ import 'package:flutter/material.dart';
 
 import '../../../zeta_flutter.dart';
 
-export './assist_chip.dart';
-export './filter_chip.dart';
-export './input_chip.dart';
-export './status_chip.dart';
+export 'assist_chip.dart';
+export 'filter_chip.dart';
+export 'input_chip.dart';
+export 'status_chip.dart';
 
 /// This covers the broad functionality of [ZetaAssistChip], [ZetaFilterChip] and [ZetaInputChip].
 ///
 /// If [selected] is not null, the chip will have the toggle behavior of [ZetaFilterChip].
+///
+/// {@category Components}
 class ZetaChip extends ZetaStatefulWidget {
   /// Constructs a [ZetaChip].
   const ZetaChip({
