+++ lib/src/components/components.dart
@@ -11,13 +11,13 @@ export 'bottom sheets/menu_items.dart';
 export 'breadcrumb/breadcrumb.dart' hide TruncatedItem;
 export 'button_group/button_group.dart';
 export 'buttons/button.dart';
-export 'buttons/button_style.dart';
+export 'buttons/button_style.dart' show ZetaButtonType;
 export 'buttons/icon_button.dart';
 export 'chat_item/chat_item.dart';
+export 'chat_item/contact_item.dart';
 export 'checkbox/checkbox.dart' hide ZetaInternalCheckbox;
 export 'chips/chip.dart';
 export 'comms_button/comms_button.dart';
-export 'contact_item/contact_item.dart';
 export 'date_input/date_input.dart';
 export 'dial_pad/dial_pad.dart';
 export 'dialog/dialog.dart';
