+++ lib/src/components/date_input/date_input.dart
@@ -5,7 +5,6 @@ import 'package:intl/intl.dart';
 import 'package:mask_text_input_formatter/mask_text_input_formatter.dart';
 
 import '../../../zeta_flutter.dart';
-import '../../interfaces/form_field.dart';
 import '../buttons/input_icon_button.dart';
 import '../text_input/internal_text_input.dart';
