+++ lib/src/components/list_item/notification_list_item.dart
@@ -267,6 +267,8 @@ extension on Image {
 
 // TODO(UX-1138): Can this be refactored to use ZetaIndicator?
 /// Badge item for notification list items. Can be an avatar, icon or image
+///
+/// {@category Components}
 class ZetaNotificationBadge extends StatelessWidget {
   /// Constructs a notification badge with an avatar.
   const ZetaNotificationBadge.avatar({
