+++ lib/src/components/search_bar/search_bar.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import '../../../zeta_flutter.dart';
-import '../../interfaces/form_field.dart';
 import '../buttons/input_icon_button.dart';
 import '../text_input/internal_text_input.dart';
