+++ lib/src/components/select_input/select_input.dart
@@ -3,7 +3,6 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 import '../../../zeta_flutter.dart';
-import '../../interfaces/form_field.dart';
 import '../buttons/input_icon_button.dart';
 import '../dropdown/dropdown_controller.dart';
 import '../text_input/internal_text_input.dart';
