+++ lib/src/components/stepper/stepper.dart
@@ -490,6 +490,8 @@ class VerticalStep extends StatelessWidget {
 /// Zeta step used in [ZetaStepper]. The step can have a title and subtitle,
 /// an icon within its circle, some content and a state that governs its
 /// styling.
+///
+/// {@category Components}
 class ZetaStep {
   /// Creates a step for a [ZetaStepper].
   const ZetaStep({
