+++ lib/src/components/time_input/time_input.dart
@@ -4,7 +4,6 @@ import 'package:flutter/services.dart';
 import 'package:mask_text_input_formatter/mask_text_input_formatter.dart';
 
 import '../../../zeta_flutter.dart';
-import '../../interfaces/form_field.dart';
 import '../buttons/input_icon_button.dart';
 import '../text_input/internal_text_input.dart';
