+++ lib/src/components/top_app_bar/search_top_app_bar.dart
@@ -183,6 +183,8 @@ class _ZetaTopAppBarSearchFieldState extends State<ZetaTopAppBarSearchField> wit
 }
 
 /// A controller used to control the search field in the app bar.
+///
+/// {@category Utils}
 class ZetaSearchController extends ChangeNotifier {
   bool _enabled = false;
