+++ lib/src/interfaces/countries.dart
@@ -1,4 +1,6 @@
-/// [Country]
+/// Class definition for a country with its name, dial code, and ISO code.
+///
+/// {@category Components}
 class Country {
   /// Constructor for [Country]
   const Country({
@@ -27,7 +29,9 @@ class Country {
       }.toString();
 }
 
-/// [Countries]
+/// Class definition for a list of countries.
+///
+/// {@category Interfaces}
 class Countries {
   /// List of [Country] for most countries around the world.
   static List<Country> get list => _countriesList;
