+++ lib/src/interfaces/form_field.dart
@@ -42,6 +42,8 @@ abstract class ZetaFormField<T> extends FormField<T> {
 }
 
 /// A text form field used in Zeta
+///
+/// {@category Interfaces}
 abstract class ZetaTextFormField extends ZetaFormField<String> {
   /// Creates a new [ZetaTextFormField]
   ZetaTextFormField({
