+++ lib/src/interfaces/phone_number.dart
@@ -1,0 +1,16 @@
+/// A phone number.
+///
+/// {@category Interfaces}
+class PhoneNumber {
+  /// Creates a new [PhoneNumber].
+  const PhoneNumber({
+    required this.dialCode,
+    required this.number,
+  });
+
+  /// The dial code of the phone number.
+  final String dialCode;
+
+  /// The number of the phone number.
+  final String number;
+}
