+++ lib/src/theme/color_swatch.dart
@@ -178,6 +178,8 @@ class ZetaColorSwatch extends ColorSwatch<int> with EquatableMixin {
 }
 
 /// A swatch of colors with values 0 (light), 500 (medium) and 1000(dark).
+///
+/// {@category Theme}
 class ZetaPureColorSwatch extends ColorSwatch<int> with EquatableMixin {
   /// Constructs a [ZetaPureColorSwatch].
   ///
