+++ lib/src/theme/custom_theme.dart
@@ -3,6 +3,8 @@ import 'package:flutter/material.dart';
 import 'color_swatch.dart';
 
 /// A custom theme that can be used to define custom colors for the app.
+///
+/// {@category Theme}
 class ZetaCustomTheme {
   /// Constructs a [ZetaCustomTheme].
   /// To define every shade of a color, provide a [ZetaColorSwatch] or a [MaterialColor].
