+++ lib/src/utils/nothing.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/widgets.dart';
 
 /// A convenient widget that renders nothing.
+///
 /// {@category Utils}
 class Nothing extends StatelessWidget {
   /// Constructs a [Nothing] widget.
