+++ lib/zeta_flutter.dart
@@ -1,7 +1,8 @@
 /// Zebra Design System (Zeta) - Flutter Component Library
 library;
 
-export 'generated/generated.dart';
 export 'src/components/components.dart';
+export 'src/generated/generated.dart';
+export 'src/interfaces/interfaces.dart';
 export 'src/theme/theme.dart';
 export 'src/utils/utils.dart';
